+++ src/store/index.ts
@@ -1,4 +1,7 @@
 export * from './getEnvironmentConfig';
+export * from './httpFile';
 export * from './httpFileStore';
+export * from './httpRegion';
+export * from './parser';
 export * from './pluginStore';
 export * from './userSessionStore';