+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [unreleased]
+### Features
+- add `--tag` support to cli to only execute httpRegion with defined tag (#693)
+
 ## [6.13.3]
 ### Fix
 - Aws Signing use query params in signing request (#684)